+++ lib/seed_ui.dart
@@ -28,6 +28,8 @@ export 'src/components/data_display/collapse.dart'
         CollapseToken,
         CollapseIconPosition,
         CollapsibleTrigger;
+export 'src/components/data_display/countdown.dart'
+    show Countdown, CountdownType, CountdownToken;
 export 'src/components/data_display/empty.dart'
     show Empty, EmptyImage, EmptyToken;
 export 'src/components/data_display/listy.dart'