+++ lib/src/theme/components_config.dart
@@ -4,6 +4,7 @@ import '../components/data_display/avatar.dart';
 import '../components/data_display/badge.dart';
 import '../components/data_display/card.dart';
 import '../components/data_display/collapse.dart';
+import '../components/data_display/countdown.dart';
 import '../components/data_display/empty.dart';
 import '../components/data_display/listy.dart';
 import '../components/data_display/popover.dart';
@@ -48,6 +49,7 @@ class ComponentsConfig {
     this.card,
     this.checkbox,
     this.collapse,
+    this.countdown,
     this.drawer,
     this.dropdown,
     this.empty,
@@ -99,6 +101,9 @@ class ComponentsConfig {
 
   /// Overrides applied to every [Collapse] under this provider.
   final CollapseToken? collapse;
+
+  /// Overrides applied to every [Countdown] under this provider.
+  final CountdownToken? countdown;
 
   /// Overrides applied to every [Drawer] under this provider.
   final DrawerToken? drawer;
@@ -196,6 +201,7 @@ class ComponentsConfig {
     if (T == CardToken && card != null) return card as T;
     if (T == CheckboxToken && checkbox != null) return checkbox as T;
     if (T == CollapseToken && collapse != null) return collapse as T;
+    if (T == CountdownToken && countdown != null) return countdown as T;
     if (T == DrawerToken && drawer != null) return drawer as T;
     if (T == DropdownToken && dropdown != null) return dropdown as T;
     if (T == EmptyToken && empty != null) return empty as T;